refactor: Moved dtos to separate package
