+++ src/main/java/school/faang/user_service/dto/recommendation/RejectionDto.java
@@ -1,0 +1,4 @@
+package school.faang.user_service.dto.recommendation;
+
+public record RejectionDto(String reason) {
+}
