--- src/main/java/school/faang/user_service/repository/recommendation/RejectionDto.java
@@ -1,4 +1,0 @@
-package school.faang.user_service.repository.recommendation;
-
-public record RejectionDto(String reason) {
-}
